Make books user specific

+++ backend/schema/resolvers.js
@@ -3,17 +3,24 @@ const Book = require("../models/Book");
 
 const resolvers = {
   Query: {
-    books: async () => {
-      const cachedBooks = await redisClient.get("books");
+    books: async (_, __, context) => {
+      
+      if(!context.user){
+        return [];
+      }
+
+      const cacheKey = `books:${context.user.uid}`;
+
+      const cachedBooks = await redisClient.get(cacheKey);
 
       if(cachedBooks){
         console.log("Books served from Redis cache");
         return JSON.parse(cachedBooks);
       }
 
-      console.log("Books served from memory");
+      console.log("User books served from MongoDB");
 
-      const books = await Book.find().sort({createdAt: -1})
+      const books = await Book.find({userId: context.user.uid}).sort({createdAt: -1})
 
       await redisClient.set("books", JSON.stringify(books), {
         EX: 60,
@@ -33,9 +40,10 @@ const resolvers = {
         title,
         author,
         notes,
+        userId: context.user.uid,
       });
 
-      await redisClient.del("books");
+      await redisClient.del(`books:${context.user.uid}`);
       return newBook;
     },
 
@@ -43,13 +51,13 @@ const resolvers = {
       if(!context.user){
         throw new Error("You must be logged in to delete books");
       }
-      const deletedBook = await Book.findByIdAndDelete(id);
+      const deletedBook = await Book.findByIdAndDelete({_id: id, userId: context.user.uid});
 
       if(!deletedBook){
-        throw new Error("Book not found")
+        throw new Error("Book not found or you are not allowed to delete it")
       }
 
-      await redisClient.del("books");
+      await redisClient.del(`books:${context.user.uid}`);
 
       return deletedBook;
     },
@@ -73,10 +81,10 @@ const resolvers = {
       );
 
       if(!updatedBook){
-        throw new Error("Book nto found");
+        throw new Error("Book not found or you are not allowed to update it");
       }
 
-      await redisClient.del("books");
+      await redisClient.del(`books:${context.user.uid}`);
 
       return updatedBook;
     }
